Revert "Get Mission detail api"
