+++ API-Server/src/main/java/swm/hkcc/LGTM/app/global/constant/ResponseCode.java
@@ -42,11 +42,7 @@ public enum ResponseCode {
     // 101xx : Member Authority
     NOT_EXIST_MEMBER(10100, HttpStatus.BAD_REQUEST, "존재하지 않는 회원입니다."),
     NOT_SENIOR_MEMBER(10101, HttpStatus.BAD_REQUEST, "시니어 회원만 접근할 수 있습니다."),
-    NOT_JUNIOR_MEMBER(10102, HttpStatus.BAD_REQUEST, "주니어 회원만 접근할 수 있습니다."),
-
-    // 102xx : Mission
-    NOT_EXIST_MISSION(10200, HttpStatus.BAD_REQUEST, "존재하지 않는 미션입니다."),
-    ;
+    NOT_JUNIOR_MEMBER(10102, HttpStatus.BAD_REQUEST, "주니어 회원만 접근할 수 있습니다.");
 
     private final Integer code;
     private final HttpStatus httpStatus;
