+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/auth/service/AuthServiceImpl.java
@@ -23,7 +23,6 @@ import swm.hkcc.LGTM.app.modules.member.domain.Senior;
 import swm.hkcc.LGTM.app.modules.member.repository.JuniorRepository;
 import swm.hkcc.LGTM.app.modules.member.repository.MemberRepository;
 import swm.hkcc.LGTM.app.modules.member.repository.SeniorRepository;
-import swm.hkcc.LGTM.app.modules.member.service.MemberService;
 import swm.hkcc.LGTM.app.modules.tag.domain.TechTag;
 import swm.hkcc.LGTM.app.modules.tag.domain.TechTagPerMember;
 import swm.hkcc.LGTM.app.modules.tag.repository.TechTagPerMemberRepository;
@@ -38,13 +37,15 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Service
 public class AuthServiceImpl implements AuthService {
+    private static final String JUNIOR = "JUNIOR";
+    private static final String SENIOR = "SENIOR";
+
     private final MemberRepository memberRepository;
     private final JuniorRepository juniorRepository;
     private final SeniorRepository seniorRepository;
     private final TechTagRepository techTagRepository;
     private final TechTagPerMemberRepository techTagPerMemberRepository;
     private final TokenProvider tokenProvider;
-    private final MemberService memberService;
 
     @Override
     @Transactional
@@ -67,7 +68,7 @@ public class AuthServiceImpl implements AuthService {
                     .accessToken(createAccessToken(updatedMember))
                     .refreshToken(refreshToken)
                     .profileImageUrl(githubUserInfo.getAvatar_url())
-                    .memberType(memberService.getMemberType(updatedMember.getMemberId()))
+                    .memberType(getMemberType(updatedMember.getMemberId()))
                     .build();
         }
 
@@ -177,8 +178,27 @@ public class AuthServiceImpl implements AuthService {
                 .githubId(member.getGithubId())
                 .accessToken(createAccessToken(member))
                 .refreshToken(member.getRefreshToken())
-                .memberType(memberService.getMemberType(member.getMemberId()))
+                .memberType(getMemberType(member.getMemberId()))
                 .build();
     }
 
+    private String getMemberType(Long memberId) {
+        if (isJunior(memberId)) {
+            return JUNIOR;
+        }
+        else if (isSenior(memberId)) {
+            return SENIOR;
+        }
+        else {
+            throw new UnspecifiedMemberType();
+        }
+    }
+
+    private boolean isJunior(Long memberId) {
+        return juniorRepository.existsByMember_MemberId(memberId);
+    }
+
+    private boolean isSenior(Long memberId) {
+        return seniorRepository.existsByMember_MemberId(memberId);
+    }
 }
