+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/member/repository/SeniorRepository.java
@@ -4,10 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import swm.hkcc.LGTM.app.modules.member.domain.Senior;
 
-import java.util.Optional;
-
 @Repository
 public interface SeniorRepository extends JpaRepository<Senior, Long> {
     boolean existsByMember_MemberId(Long memberId);
-    Optional<Senior> findByMember_MemberId(Long memberId);
 }
