+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/mission/controller/CreateMissionController.java
@@ -3,22 +3,22 @@ package swm.hkcc.LGTM.app.modules.mission.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import swm.hkcc.LGTM.app.global.dto.ApiDataResponse;
 import swm.hkcc.LGTM.app.modules.member.domain.custom.CustomUserDetails;
 import swm.hkcc.LGTM.app.modules.mission.domain.Mission;
 import swm.hkcc.LGTM.app.modules.mission.dto.CreateMissionRequest;
 import swm.hkcc.LGTM.app.modules.mission.dto.CreateMissionResponse;
-import swm.hkcc.LGTM.app.modules.mission.dto.MissionDetailViewResponse;
 import swm.hkcc.LGTM.app.modules.mission.service.CreateMissionService;
-import swm.hkcc.LGTM.app.modules.mission.service.MissionService;
 
 @RestController
 @RequestMapping("/v1/mission")
 @RequiredArgsConstructor
-public class MissionController {
+public class CreateMissionController {
     private final CreateMissionService createMissionService;
-    private final MissionService missionService;
 
     @PostMapping
     public ApiDataResponse<CreateMissionResponse> createMissinon(
@@ -33,15 +33,5 @@ public class MissionController {
                 .missionId(mission.getMissionId())
                 .writerId(memberId)
                 .build());
-    }
-
-
-    @GetMapping("/detail")
-    public ApiDataResponse<MissionDetailViewResponse> getMission(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails,
-            @RequestParam Long missionId
-    ) {
-        Long memberId = customUserDetails.getMemberId();
-        return ApiDataResponse.of(missionService.getMissionDetail(memberId, missionId));
     }
 }
