--- API-Server/src/main/java/swm/hkcc/LGTM/app/modules/mission/exception/NotExistMission.java
@@ -1,10 +1,0 @@
-package swm.hkcc.LGTM.app.modules.mission.exception;
-
-import swm.hkcc.LGTM.app.global.constant.ResponseCode;
-import swm.hkcc.LGTM.app.global.exception.GeneralException;
-
-public class NotExistMission extends GeneralException {
-    public NotExistMission() {
-        super(ResponseCode.NOT_EXIST_MISSION);
-    }
-}
