+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/mission/service/MissionService.java
@@ -1,7 +1,6 @@
 package swm.hkcc.LGTM.app.modules.mission.service;
 
 import swm.hkcc.LGTM.app.modules.mission.domain.MissionContentData;
-import swm.hkcc.LGTM.app.modules.mission.dto.MissionDetailViewResponse;
 
 public interface MissionService {
     MissionContentData getOngoingMissions(Long memberId);
@@ -9,6 +8,4 @@ public interface MissionService {
     MissionContentData getRecommendMissions(Long memberId);
 
     MissionContentData getTotalMissions(Long memberId);
-
-    MissionDetailViewResponse getMissionDetail(Long memberId, Long missionId);
 }
