+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/mission/service/MissionServiceImpl.java
@@ -3,17 +3,10 @@ package swm.hkcc.LGTM.app.modules.mission.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import swm.hkcc.LGTM.app.modules.member.domain.Senior;
-import swm.hkcc.LGTM.app.modules.member.exception.NotExistMember;
-import swm.hkcc.LGTM.app.modules.member.repository.MemberRepository;
-import swm.hkcc.LGTM.app.modules.member.repository.SeniorRepository;
-import swm.hkcc.LGTM.app.modules.member.service.MemberService;
 import swm.hkcc.LGTM.app.modules.mission.domain.Mission;
 import swm.hkcc.LGTM.app.modules.mission.domain.MissionContentData;
 import swm.hkcc.LGTM.app.modules.mission.domain.mapper.MissionMapper;
-import swm.hkcc.LGTM.app.modules.mission.dto.MissionDetailViewResponse;
 import swm.hkcc.LGTM.app.modules.mission.dto.MissionDetailsDto;
-import swm.hkcc.LGTM.app.modules.mission.exception.NotExistMission;
 import swm.hkcc.LGTM.app.modules.mission.repository.MissionRepository;
 import swm.hkcc.LGTM.app.modules.mission.repository.MissionScrapRepository;
 import swm.hkcc.LGTM.app.modules.mission.repository.MissionViewRepository;
@@ -23,8 +16,6 @@ import swm.hkcc.LGTM.app.modules.tag.repository.TechTagPerMissionRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
-import static swm.hkcc.LGTM.app.modules.mission.domain.mapper.MissionMapper.missionAndMemberToDetailView;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -36,9 +27,6 @@ public class MissionServiceImpl implements MissionService {
     private final MissionViewRepository missionViewRepository;
     private final MissionRegistrationRepository missionRegistrationRepository;
     private final TechTagPerMissionRepository techTagPerMissionRepository;
-    private final MemberRepository memberRepository;
-    private final SeniorRepository seniorRepository;
-    private final MemberService memberService;
 
     @Override
     public MissionContentData getOngoingMissions(Long memberId) {
@@ -78,21 +66,6 @@ public class MissionServiceImpl implements MissionService {
                 .collect(Collectors.toList());
 
         return MissionContentData.of(missionDetailsDtos);
-    }
-
-    @Override
-    public MissionDetailViewResponse getMissionDetail(Long memberId, Long missionId) {
-        Mission mission = missionRepository.findById(missionId).orElseThrow(NotExistMission::new);
-        boolean isScraped = missionScrapRepository.existsByScrapper_MemberIdAndMission_MissionId(memberId, missionId);
-
-        Long missionWriterId = mission.getWriter().getMemberId();
-        Senior missionWriter = seniorRepository.findByMember_MemberId(missionWriterId).orElseThrow(NotExistMember::new);
-
-        List<TechTag> techTagList = techTagPerMissionRepository.findTechTagsByMissionId(mission.getMissionId());
-        int currentPeopleNumber = missionRegistrationRepository.countByMission_MissionId(mission.getMissionId());
-        String memberType = memberService.getMemberType(memberId);
-
-        return missionAndMemberToDetailView(mission, isScraped, missionWriter, techTagList, currentPeopleNumber, memberType);
     }
 
     // 미완성
